Move string comparison utility to the generic utils class.
