--- tests/commands/convert/utils.py
@@ -1,8 +1,0 @@
-class PartialMatchString(str):
-    def __eq__(self, other):
-        return self in other
-
-
-class NoMatchString(str):
-    def __eq__(self, other):
-        return self not in other
